Move modules separate to script
Deployment failed, so I've moved the helper functions out separate to the script in the hope it'll work better
