+++ scripts/seed-staging.py
@@ -1,4 +1,10 @@
-from .seed import commit_data, clear_old_data
+from modules.seed import commit_data, clear_old_data
 
-clear_old_data()
-commit_data()
+
+def main():
+    clear_old_data()
+    commit_data()
+
+
+if __name__ == '__main__':
+    main()
